IU0QWJ Edition v4.3.1: 999 channels, EEPROM overlay system, all features
Hardware: 24M02 EEPROM (256KB) replacing original 24C64 (8KB).

Firmware changes:
- 999 MR channels (up from 200) with new EEPROM memory map
- EEPROM code overlay system: Spectrum, FM Radio, Aircopy, and Menu
  rendering loaded from EEPROM into RAM on demand, overcoming the 60KB
  flash limit and enabling all features in a single build
- All features enabled: Spectrum (F+6), NOAA (F+5), FM Radio, Aircopy,
  VOX, TX1750, Flashlight, DTMF Calling, Charging C
- Factory calibration preserved at original addresses (0x1E00-0x1FFF)
- CHIRP driver updated for 999 channels + overlay binary upload
- Docker build script (compile-iu0qwj.sh) and CI workflow updated

New files: overlay.c/h, eeprom_map.h, ui/menu_render.c, compile-iu0qwj.sh

Made-with: Cursor

+++ driver/eeprom.h
@@ -19,8 +19,7 @@
 
 #include <stdint.h>
 
-void EEPROM_ReadBuffer(uint16_t Address, void *pBuffer, uint8_t Size);
-void EEPROM_WriteBuffer(uint16_t Address, const void *pBuffer);
+void EEPROM_ReadBuffer(uint32_t Address, void *pBuffer, uint16_t Size);
+void EEPROM_WriteBuffer(uint32_t Address, const void *pBuffer);
 
 #endif
-
